+++ app/controllers/starter_file_groups_controller.rb
@@ -6,7 +6,11 @@ class StarterFileGroupsController < ApplicationController
 
   def create
     assignment = Assignment.find_by(id: params[:assignment_id])
-    assignment.starter_file_groups.create(update_params)
+
+    new_params = update_params
+    new_params[:name] ||= StarterFileGroup.autogenerate_starter_file_group_name(assignment)
+
+    assignment.starter_file_groups.create(new_params)
   end
 
   def destroy