+++ Changelog.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 
 - Fix bug in displaying feedback files for test results (#5719)
+- Require starter group names to be unique for an assignment (#5721)
 
 ## [v2.0.1]
 - Fix bug where a login with remote authentication failed to redirect to the landing page (#5690)